requirements.txt
requests
pyyaml
pandas
pyarrow
streamlit
